Remove PaymentKeeperSubscriber.php Reworked IDealIssuerSubscriber.php to be a ProcessPaymentDisplaySubscriber.php Updated CHANGELOG.md and README.md to be a 1.1.0 version
