+++ custom/plugins/emspay/CHANGELOG.md
@@ -8,3 +8,9 @@
 
 * Removed unnecessary typecasts for order lines
 * Label and translations changes in configuration file
+
+** 1.1.0 **
+
+* Merged a logic from IDealIssuerSubscriber and PaymentKeeper to be a single ProcessPaymentDisplaySubscriber.
+* Added iBAN information for BankTransfer payment method.
+* Implemented country validation feature for Afterpay payment method.
