+++ custom/plugins/emspay/Resources/services.xml
@@ -15,13 +15,10 @@
         <service id="emspay_events.subscriber.checkout_subscriber" class="emspay\Subscriber\CheckoutSubscriber">
             <tag name="shopware.event_subscriber"/>
         </service>
-        <service id="emspay_events.subscriber.idealissuer_subscriber" class="emspay\Subscriber\IDealIssuerSubscriber">
+        <service id="emspay_events.subscriber.processpaymentdispaly_subscriber" class="emspay\Subscriber\ProcessPaymentDisplaySubscriber">
             <tag name="shopware.event_subscriber"/>
         </service>
         <service id="emspay_events.subscriber.afterpay_subscriber" class="emspay\Subscriber\AfterpaySubscriber">
-            <tag name="shopware.event_subscriber"/>
-        </service>
-        <service id="emspay_events.subscriber.paymentkeeper_subscriber" class="emspay\Subscriber\PaymentKeeperSubscriber">
             <tag name="shopware.event_subscriber"/>
         </service>
     </services>
