+++ custom/plugins/emspay/plugin.xml
@@ -4,7 +4,7 @@
 
     <label >EMS Online Ginger Payments</label>
 
-    <version>1.0.1</version>
+    <version>1.1.0</version>
     <copyright>(c) 2020 EMS</copyright>
     <author>Ginger Payments</author>
